+++ TODO.md
@@ -1,6 +1,8 @@
 
 # Ideas to ponder
 
+- rename LogLine/Line to Entry?
+
 - Consider replacing SeedModifier with a seed modifier type...
 
   ```
@@ -25,6 +27,19 @@
 - Allow base loggers to share resources (span attribute tracking, JSON formatting)
 
 - Figure out a way to enforce limited vocabulary on select Attributes and Data elements
+
+- Some way to mix-in data from code context.  The log object will be in the call stack
+  either passed explicityly or passed within context.  If the call is to a method on an
+  initialized library, that library may want to wrap logging to:
+
+  - adjust the logging level that is being retained
+  - add it's own line prefix attributes
+
+  What's the right API for this?
+
+- Logging levels.  Are they actually useful?  Let's pretend they are.  What's the right
+  behavior.  Baseloggers could choose to ignore levels that are too low.  The logger itself
+  could choose to discard when the level is too low.
 
 # Just do it (build ready)
 
@@ -96,6 +111,16 @@
 
   - Gateway into Jaeger
   - Gateway into Open Telementry
+  - Improve the JSON logger
+
+    It can have a bunch more knobs and controls
+
+    - add additional time formats
+    - add ability to maintain and send dictionaries to compress logs
+
+      - keys for key/value
+      - span ids (use the spanSequenceCode)
+      - log line prefix sets
 
 - Bytes writers:
 
@@ -111,4 +136,5 @@
   - Gets counted
   - Re-use attributes?
   - Attach arbitrary data
+
